requirements.txt
discord.py==2.4.0
matplotlib==3.9.0
pandas==2.2.2
pybit==5.7.0
python-dotenv==1.0.1
ta==0.11.0
pycryptodome~=3.20.0